+++ src/e_help.tcl
@@ -11,10 +11,7 @@
 #
 # Local help is downloaded with wget:
 # wget -r -k -l 2 -p --accept-regex=.+/man/tcl8\.6.+ https://www.tcl.tk/man/tcl8.6/
-#
-# *******************************************************************
-# Scripted by Alex Plotnikov
-# *******************************************************************
+# _______________________________________________________________________ #
 
 package require Tk
 
@@ -39,24 +36,8 @@ namespace eval ::eh {
   variable solo [expr {[info exist ::argv0] && [file normalize $::argv0] eq \
     [file normalize [info script]]} ? 1 : 0]
 }
-
-# *******************************************************************
-# common procedures
-
-#====== debug messages (several params)
-proc ddd {args} {
-  set msg ""; foreach l $args {append msg " $l\n"}
-  ::apave::obj ok info DEBUG $msg -text 1 -h 10 -centerme .
-}
-
-#====== debug messages (array passed by name)
-proc a {a} {set m [array get $a]; d $m}
-
-# *******************************************************************
-# e_help's procedures
 #=== own message/question box
 proc ::eh::dialog_box {ttl mes {typ ok} {icon info} {defb OK} args} {
-  set pdlg [::apave::APaveDialog new]
   set opts [list -t 1 -w 80]
   lappend opts {*}$args
   switch -glob -- $typ {
@@ -64,22 +45,12 @@ proc ::eh::dialog_box {ttl mes {typ ok} {icon info} {defb OK} args} {
       if {$defb eq "OK" && $typ ne "okcancel" } {
         set defb YES
       }
-      set ans [$pdlg $typ $icon $ttl \n$mes\n $defb {*}$opts]
+      set ans [::apave::obj $typ $icon $ttl \n$mes\n $defb {*}$opts]
     }
     default {
-      set ans [$pdlg ok $icon $ttl \n$mes\n {*}$opts]
+      set ans [::apave::obj ok $icon $ttl \n$mes\n {*}$opts]
     }
   }
-  $pdlg destroy
-  return $ans
-}
-#====== 'mes' message of 'typ' type
-proc ::eh::message_box {mes {typ ok} {ttl ""}} {
-  if {[string length $ttl] == 0} {set ttl [wm title .]}
-  set mes [string trimleft $mes "\{"]
-  set mes [string trimright $mes "\}"]
-  set ans [tk_messageBox -title $ttl -icon info -message "$mes" \
-    -type $typ -parent .]
   return $ans
 }
 #=== get terminal's name
@@ -90,7 +61,7 @@ proc ::eh::get_tty {inconsole} {
   else {set tty xterm}
   return $tty
 }
-#====== to get system time & date
+#=== get system time & date
 proc ::eh::get_timedate {} {
   set systime [clock seconds]
   set curtime [clock format $systime -format $::eh::formtime]
@@ -106,8 +77,7 @@ proc ::eh::get_language {} {
   }
   return $lang
 }
-
-#====== to maximize 'win' window
+#=== maximize 'win' window
 proc ::eh::zoom_window {win} {
   if {[::iswindows]} {
     wm state $win zoomed
@@ -115,7 +85,7 @@ proc ::eh::zoom_window {win} {
     wm attributes $win -zoomed 1
   }
 }
-#====== to center window on screen
+#=== center window on screen
 proc ::eh::center_window {win {ornament 1} {winwidth 0} {winheight 0}} {
   # to center a window regarding taskbar(s) sizes
   #  center_window win     ;# if win window has borders and titlebar
@@ -169,7 +139,6 @@ proc ::eh::checkgeometry {} {
     wm geometry . ${w}x${h}+${x}+${y}
   }
 }
-
 #=== off ctrl/alt modificators of keystrokes
 proc ::eh::ctrl_alt_off {cmd} {
   if {[::iswindows]} {
@@ -178,12 +147,10 @@ proc ::eh::ctrl_alt_off {cmd} {
     return "if \{\[expr %s&14\] == 0\} \{$cmd\}"
   }
 }
-
 #=== try and check if 'app' destroyed
 proc ::eh::destroyed {app} {
   return [expr ![catch {send -async $app {destroy .}} e]]
 }
-
 #=== drag window by snatching header
 proc ::eh::mouse_drag {win mode x y} {
   switch -- $mode {
@@ -198,7 +165,6 @@ proc ::eh::mouse_drag {win mode x y} {
     }
   }
 }
-
 #=== Gets/sets file attributes
 proc ::eh::fileAttributes {fname {attrs "-"} {atime ""} {mtime ""} } {
     if {$attrs eq "-"} {
@@ -212,7 +178,6 @@ proc ::eh::fileAttributes {fname {attrs "-"} {atime ""} {mtime ""} } {
      file mtime $fname $mtime
    }
 }
-
 #=== Write data to a file with file attributes untouched
 proc ::eh::write_file_untouched {fname data} {
   lassign [::eh::fileAttributes $fname] f_attrs f_atime f_mtime
@@ -222,7 +187,6 @@ proc ::eh::write_file_untouched {fname data} {
   close $ch
   ::eh::fileAttributes $fname $f_attrs $f_atime $f_mtime
 }
-
 #=== escape double quotes
 proc ::eh::escape_quotes {sel} {
   if {![::iswindows]} {
@@ -230,31 +194,26 @@ proc ::eh::escape_quotes {sel} {
   }
   return $sel
 }
-
 #=== escape special characters
 proc ::eh::escape_specials {sel} {
   return [string map [ list \" \\\" "\n" "\\n" "\\" "\\\\" "\$" "\\\$" \
     "\}" "\\\}"  "\{" "\\\{"  "\]" "\\\]"  "\[" "\\\[" ] $sel]
 }
-
 #=== prepare "search links" for browser
 proc ::eh::escape_links {sel} {
   return [string map [list " " "+"] $sel]
 }
-
 #=== delete specials & underscore spaces
 proc ::eh::delete_specsyms {sel {und "_"} } {
   return [string map [list \
       "\"" ""  "\%" ""  "\$" ""  "\}" ""  "\{" "" \
       "\]" ""  "\[" ""  "\>" ""  "\<" ""  "\*" ""  " " $und] $sel]
 }
-
 #=== get "underlined" name (e.g. working dir)
 proc ::eh::get_underlined_name {name} {
   return [string map {/ _ \\ _ { } _ . _} $name]
 }
-
-#====== check if link exists
+#=== check if link exists
 proc ::eh::lexists {url} {
   if {$::eh::reginit} {
     set ::eh::reginit 0
@@ -263,9 +222,7 @@ proc ::eh::lexists {url} {
     ::http::register https 443 ::tls::socket
   }
   if {[catch {set token [::http::geturl $url]} e]} {
-    if {$::eh::solo} {
-    grid [label .l -text ""]}  ;# hide wish
-    message_box "ERROR: couldn't connect to:\n\n$url\n\n$e"
+    tk_messageBox -message "ERROR: couldn't connect to:\n\n$url\n\n$e"
     return 0
   }
   if {$::eh::solo} { exit }
@@ -275,7 +232,7 @@ proc ::eh::lexists {url} {
     return 0
   }
 }
-#====== check if links exist
+#=== check if links exist
 proc ::eh::links_exist {h1 h2 h3} {
   if {[lexists "$h1"]} {
     return "$h1"            ;# Tcl commands help
@@ -287,7 +244,7 @@ proc ::eh::links_exist {h1 h2 h3} {
     return ""
   }
 }
-#====== offline help
+#=== offline help
 proc ::eh::local { {help ""} } {
   set l1 [string toupper [string range "$help" 0 0]]
   if {[string first "http" "$::eh::hroot"]==0} {
@@ -318,7 +275,7 @@ proc ::eh::local { {help ""} } {
   }
   return "$h1"
 }
-#====== online help, change links if need
+#=== online help, change links if need
 proc ::eh::html { {help ""} {local 0}} {
   if {$local} {
     return [local "$help"]
@@ -333,7 +290,7 @@ proc ::eh::html { {help ""} {local 0}} {
   }
   return $link
 }
-#====== to call browser
+#=== call browser
 proc ::eh::browse { {help ""} } {
   if {$::eh::my_browser ne ""} {
     # my_browser may contain options, e.g. "chromium --no-sandbox" for root
@@ -342,8 +299,7 @@ proc ::eh::browse { {help ""} } {
     ::apave::openDoc "$help"
   }
 }
-
-# *******************************************************************
+# _______________________________________________________________________ #
 
 if {$::eh::solo} {
   if {$argc > 0} {
@@ -354,7 +310,8 @@ if {$::eh::solo} {
     }
     ::eh::browse "$page"
   } else {
-    puts "\nRun:
+    puts "
+Run:
 
   tclsh e_help.tcl \[-local\] page
 
@@ -362,8 +319,8 @@ to get Tcl/Tk help page:
 
   TclCmd/page.htm or
   TkCmd/page.htm or
-  Keywords/P.htm\n"
-  }
+  Keywords/P.htm
+"}
   exit
 }
-# *****************************   EOF   *****************************
+# ________________________________  EOF _________________________________ #